move files to comply with existing structure
