+++ internal/pkg/defiscrapers/yearn/APROracle.abi
