+++ internal/pkg/defiscrapers/yearn/APROracle.bin
