+++ internal/pkg/defiscrapers/yearn/gen.sh
